CB-20060 Move LoadBalancerConfigConverter to a separate package
It will be refactored soon, it is a preparation.
